requirements.txt
# Core requirements for SeedVR (excluding torch which is installed separately)
einops==0.7.0
omegaconf==2.3.0
opencv-python==4.9.0.80
diffusers==0.29.1
rotary-embedding-torch==0.5.3
transformers==4.38.2
mediapy==1.2.0

# Additional dependencies from environment.yml (filtered for relevance)
accelerate==0.33.0
huggingface-hub==0.32.2
safetensors==0.4.3
numpy==1.24.4
pillow==10.3.0
requests==2.32.3
tqdm==4.66.4
psutil==6.0.0
pyyaml==6.0.1

# Gradio and web interface
gradio==4.44.0
matplotlib==3.9.0

# Video processing
ffmpeg-python==0.2.0
imageio==2.34.0
imageio-ffmpeg==0.5.1

# Additional utilities
rich==13.7.1
click==8.1.7